refactor: 게시글 관련 명령-조회 분리 (#95) - EventLikeService 분리 - EventLikeRepository는 EventLikeService에서만 사용
